refactor(test): remove helpers subfolders and flatten helper files into test directory (#246)
