+++ packages/tools/test/integration/process-execution.integration.test.ts
@@ -1,7 +1,7 @@
 import { describe, expect, test, beforeEach } from 'bun:test'
 
 import { BashTool } from '../../src/bash'
-import { createMockContext } from '../helpers/mock-context'
+import { createMockContext } from '../mock-context'
 
 describe('Process Execution Tools Integration', () => {
     let context: any
