+++ packages/tui/test/test-providers.tsx
@@ -1,9 +1,9 @@
 import { render } from 'ink-testing-library'
 import React from 'react'
 
-import { DialogProvider } from '../../src/providers/dialog'
-import { KeyboardLayerProvider } from '../../src/providers/keyboard-layer'
-import { ToastProvider } from '../../src/providers/toast'
+import { DialogProvider } from '../src/providers/dialog'
+import { KeyboardLayerProvider } from '../src/providers/keyboard-layer'
+import { ToastProvider } from '../src/providers/toast'
 
 export function renderWithProviders(ui: React.ReactNode) {
     return render(
